Refactor task_manager package
