+++ task_manager/scripts/egpsr_task_manager.py
@@ -15,15 +15,15 @@ from collections import deque
 import rclpy
 from ament_index_python.packages import get_package_share_directory
 from rclpy.node import Node
-from subtask_managers.gpsr_single_tasks import GPSRSingleTask
-from subtask_managers.gpsr_tasks import GPSRTask
+from task_manager.subtask_managers.gpsr_single_tasks import GPSRSingleTask
+from task_manager.subtask_managers.gpsr_tasks import GPSRTask
 
-# from subtask_managers.gpsr_test_commands import get_gpsr_comands
-from utils.exploration_planner import ExplorationPlanner
-from utils.baml_client.types import CommandListLLM
-from utils.logger import Logger
-from utils.status import Status
-from utils.subtask_manager import SubtaskManager, Task
+# from task_manager.subtask_managers.gpsr_test_commands import get_gpsr_comands
+from task_manager.utils.exploration_planner import ExplorationPlanner
+from task_manager.utils.baml_client.types import CommandListLLM
+from task_manager.utils.logger import Logger
+from task_manager.utils.status import Status
+from task_manager.utils.subtask_manager import SubtaskManager, Task
 
 ATTEMPT_LIMIT = 3
 MAX_COMMANDS = 3
