+++ task_manager/scripts/misc/_storing_groceries.py
@@ -6,9 +6,9 @@ Task Manager for Storing Groceries task of Robocup @Home 2025
 
 import rclpy
 from rclpy.node import Node
-from utils.logger import Logger
-from utils.status import Status
-from utils.subtask_manager import SubtaskManager, Task
+from task_manager.utils.logger import Logger
+from task_manager.utils.status import Status
+from task_manager.utils.subtask_manager import SubtaskManager, Task
 
 ATTEMPT_LIMIT = 3
