+++ task_manager/scripts/misc/receptionist_task_manager.py
@@ -10,9 +10,9 @@ from datetime import datetime
 import rclpy
 from frida_constants.vision_constants import FACE_RECOGNITION_IMAGE, IMAGE_TOPIC_RECEPTIONIST
 from rclpy.node import Node
-from utils.logger import Logger
-from utils.status import Status
-from utils.subtask_manager import SubtaskManager, Task
+from task_manager.utils.logger import Logger
+from task_manager.utils.status import Status
+from task_manager.utils.subtask_manager import SubtaskManager, Task
 
 ATTEMPT_LIMIT = 3
