+++ task_manager/scripts/test/test_manipulation_manager.py
@@ -2,9 +2,9 @@
 
 import rclpy
 from rclpy.node import Node
-from utils.status import Status
-from utils.task import Task
-from subtask_managers.manipulation_tasks import ManipulationTasks
+from task_manager.utils.status import Status
+from task_manager.utils.task import Task
+from task_manager.subtask_managers.manipulation_tasks import ManipulationTasks
 
 
 class TestManipulationManager(Node):
