+++ task_manager/task_manager/subtask_managers/hri_hand.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 
-from utils.status import Status
+from task_manager.utils.status import Status
 
-# from subtask_managers.hri_tasks import HRITasks
+# from task_manager.subtask_managers.hri_tasks import HRITasks
 
 
 class HRIHand:
