+++ task_manager/task_manager/subtask_managers/vision_tasks.py
@@ -59,10 +59,10 @@ from rclpy.action import ActionClient
 from rclpy.node import Node
 from std_msgs.msg import String
 from std_srvs.srv import SetBool, Trigger
-from utils.decorators import mockable, service_check
-from utils.logger import Logger
-from utils.status import Status
-from utils.task import Task
+from task_manager.utils.decorators import mockable, service_check
+from task_manager.utils.logger import Logger
+from task_manager.utils.status import Status
+from task_manager.utils.task import Task
 
 TIMEOUT = 8.0
 TIMEOUT_WAIT_FOR_SERVICE = 1.0
